requirements.txt
# Discord Bot Dependencies
# Install with: pip install -r requirements.txt

discord.py>=2.3.0
python-dotenv>=1.0.0

# Package Monitor Dependencies
opencv-python
requests
seeed-python-dht
grove.py